OCR_PROJECT/requirements.txt
Flask==2.3.2
Pillow==9.4.0
pytesseract==0.3.10
